First commit for calculate_volatility_wages.py
File that calculate the volatility of wage for different labor income deciles.
Wage deciles are trended (divergence between the upper and the lower deciles). 
To make the process stationary, I try two detrending procedures:
1. HP filter the data 
2. Remove a linear trend
This files produces graphs on the wage deciles and a table with the volatility of wage deciles.